fixing update group by remove layout
