+++ src/app/chat/page.tsx
@@ -39,12 +39,17 @@ export default function ChatRoom() {
     }, [roomId])
 
     // Socket Initialization
-    const { socket, allUsers, socketError, clearSocketError, messages, currentGroupUser } =
-        useSocket(data?.idToken)
+    const {
+        socket,
+        allUsers,
+        socketError,
+        clearSocketError,
+        messages,
+        currentGroupUser,
+    } = useSocket(data?.idToken)
 
     // Initialize sticker functionality
     const activeRoomData = chatRooms.find((r) => r.cid === activeRoom)
-
     // Debug log
     React.useEffect(() => {
         console.log(
@@ -76,8 +81,38 @@ export default function ChatRoom() {
     // Get current user and other users
     const { currentUser, otherUsers } = useCurrentUser(allUsers, data?.user)
 
+    const handleCreateDM = useCreateDM(
+        socket,
+        chatRooms,
+        setActiveRoom,
+        currentUser,
+        refreshChatRooms
+    )
+
+    const filteredRooms = chatRooms.filter((room) =>
+        room.name.toLowerCase().includes(searchTerm.toLowerCase())
+    )
+
+    const groupRooms = filteredRooms.filter((r) => r.is_groupchat === true)
+    const privateRooms = filteredRooms.filter((r) => r.is_groupchat === false)
+
+    // Handle room selection within chat page
+    const handleRoomSelect = (roomId: number) => {
+        if (socket) {
+            socket?.emit("join_chat", roomId)
+        }
+        setActiveRoom(roomId)
+        router.push(`/chat?roomId=${roomId}`)
+    }
+
+    React.useEffect(() => {
+        if (socket && activeRoom) {
+            socket.emit("join_chat", activeRoom)
+        }
+    }, [socket, activeRoom])
+
     return (
-        <div className="h-full flex bg-purple-200 gap-4 p-4">
+        <div className="h-screen flex bg-purple-200 gap-4 p-4">
             <YOTTLoading show={!isInited} />
 
             {/* Error Notification */}
@@ -94,6 +129,18 @@ export default function ChatRoom() {
                     </div>
                 </div>
             )}
+
+            {/* Sidebar - merged from AppLayoutWrapper */}
+            <Sidebar
+                searchTerm={searchTerm}
+                setSearchTerm={setSearchTerm}
+                showCreateDM={showCreateDM}
+                setShowCreateDM={setShowCreateDM}
+                groupRooms={groupRooms}
+                privateRooms={privateRooms}
+                activeRoom={activeRoom}
+                setActiveRoom={handleRoomSelect}
+            />
 
             {/* Main Content Area */}
             <div className="flex-1 flex gap-4">
@@ -117,6 +164,24 @@ export default function ChatRoom() {
                     currentGroupUser={currentGroupUser}
                 />
             </div>
+
+            {/* Create DM Modal - merged from AppLayoutWrapper */}
+            <CreateDMModal
+                isOpen={showCreateDM}
+                onClose={() => setShowCreateDM(false)}
+                onCreateDM={handleCreateDM}
+                allUsers={otherUsers}
+            />
+
+            {/* Sticker Modal - merged from AppLayoutWrapper */}
+            <StickerModal
+                isOpen={showStickerModal}
+                onClose={closeStickerModal}
+                onSelectSticker={handleSelectSticker}
+                stickerPacks={stickerPacks}
+                selectedPack={selectedPack}
+                onPackChange={setSelectedPack}
+            />
         </div>
     )
 }
